test: promeut TestProvider partagé dans @/test
